+++ tests/logging_test.py
@@ -35,7 +35,7 @@ class ActorLoggingTest(object):
         ActorRegistry.stop_all()
 
     def test_unexpected_messages_are_logged(self):
-        self.actor_ref.send_request_reply({'unhandled': 'message'})
+        self.actor_ref.tell({'unhandled': 'message'})
         self.assertEqual(1, len(self.log_handler.messages['warning']))
         log_record = self.log_handler.messages['warning'][0]
         self.assertEqual('Unexpected message received by %s' % self.actor_ref,
@@ -56,7 +56,7 @@ class ActorLoggingTest(object):
 
     def test_exception_is_logged_when_not_reply_requested(self):
         self.on_failure_was_called.clear()
-        self.actor_ref.send_one_way({'command': 'raise exception'})
+        self.actor_ref.tell({'command': 'raise exception'})
         self.assertTrue(self.on_failure_was_called.wait(5))
         self.assertEqual(1, len(self.log_handler.messages['error']))
         log_record = self.log_handler.messages['error'][0]
@@ -68,7 +68,7 @@ class ActorLoggingTest(object):
     def test_base_exception_is_logged(self):
         self.log_handler.reset()
         self.on_stop_was_called.clear()
-        self.actor_ref.send_one_way({'command': 'raise base exception'})
+        self.actor_ref.tell({'command': 'raise base exception'})
         self.assertTrue(self.on_stop_was_called.wait(5))
         self.assertEqual(3, len(self.log_handler.messages['debug']))
         log_record = self.log_handler.messages['debug'][0]