+++ tests/ref_test.py
@@ -65,6 +65,19 @@ class RefTest(object):
         except ActorDeadError:
             self.fail('Should never raise ActorDeadError')
 
+    def test_tell_delivers_message_to_actors_custom_on_receive(self):
+        self.ref.tell({'command': 'a custom message'})
+        self.assertEqual({'command': 'a custom message'},
+            self.received_message.get())
+
+    def test_tell_fails_if_actor_is_stopped(self):
+        self.ref.stop()
+        try:
+            self.ref.tell({'command': 'a custom message'})
+            self.fail('Should raise ActorDeadError')
+        except ActorDeadError as exception:
+            self.assertEqual('%s not found' % self.ref, str(exception))
+
     def test_send_one_way_delivers_message_to_actors_custom_on_receive(self):
         self.ref.send_one_way({'command': 'a custom message'})
         self.assertEqual({'command': 'a custom message'},
@@ -74,6 +87,29 @@ class RefTest(object):
         self.ref.stop()
         try:
             self.ref.send_one_way({'command': 'a custom message'})
+            self.fail('Should raise ActorDeadError')
+        except ActorDeadError as exception:
+            self.assertEqual('%s not found' % self.ref, str(exception))
+
+    def test_ask_blocks_until_response_arrives(self):
+        result = self.ref.ask({'command': 'ping'})
+        self.assertEqual('pong', result)
+
+    def test_ask_can_timeout_if_blocked_too_long(self):
+        try:
+            self.ref.ask({'command': 'ping'}, timeout=0)
+            self.fail('Should raise Timeout exception')
+        except Timeout:
+            pass
+
+    def test_ask_can_return_future_instead_of_blocking(self):
+        future = self.ref.ask({'command': 'ping'}, block=False)
+        self.assertEqual('pong', future.get())
+
+    def test_ask_fails_if_actor_is_stopped(self):
+        self.ref.stop()
+        try:
+            self.ref.ask({'command': 'ping'})
             self.fail('Should raise ActorDeadError')
         except ActorDeadError as exception:
             self.assertEqual('%s not found' % self.ref, str(exception))